+++ plasmapy/examples/plot_particle_stepper.py
@@ -11,7 +11,7 @@ import numpy as np
 from astropy import units as u
 from plasmapy.classes import Plasma
 from plasmapy.simulation import ParticleTracker
-from plasmapy.physics.parameters import gyrofrequency
+from plasmapy.formulary.physics import gyrofrequency
 
 ############################################################
 # Initialize a plasma. This will be a source of electric and magnetic
