+++ plasmapy/formulary/mathematics/mathematics.py
