+++ plasmapy/formulary/mathematics/tests/test_Fermi_integral.py
