+++ plasmapy/formulary/mathematics/tests/test_dispersion.py
