+++ plasmapy/formulary/physics/magnetostatics.py
