+++ plasmapy/formulary/physics/quantum.py
@@ -15,9 +15,10 @@ __all__ = [
 import numpy as np
 
 from astropy import units as u
-from plasmapy import (atomic, utils, mathematics)
+from plasmapy import (atomic, utils)
+from plasmapy.formulary import mathematics
 from plasmapy.utils.decorators.checks import check_quantity
-from plasmapy.physics.relativity import Lorentz_factor
+from plasmapy.formulary.physics.relativity import Lorentz_factor
 from astropy.constants.si import (c, h, hbar, m_e, eps0, e, k_B)
 
 try:
