+++ plasmapy/formulary/physics/tests/test_magnetostatics.py
@@ -2,13 +2,11 @@ import numpy as np
 import pytest
 from astropy import units as u, constants
 
-from plasmapy.physics.magnetostatics import (MagnetoStatics,
-                                             MagneticDipole,
-                                             Wire,
-                                             GeneralWire,
-                                             FiniteStraightWire,
-                                             InfiniteStraightWire,
-                                             CircularWire)
+from plasmapy.formulary.physics.magnetostatics import (MagneticDipole,
+                                                       GeneralWire,
+                                                       FiniteStraightWire,
+                                                       InfiniteStraightWire,
+                                                       CircularWire)
 
 
 mu0_4pi = constants.mu0/4/np.pi
