+++ plasmapy/formulary/transport/braginskii.py
@@ -138,10 +138,10 @@ from astropy import units as u
 from plasmapy import (atomic, utils)
 from plasmapy.atomic.atomic import _is_electron
 from astropy.constants.si import (e, m_e, k_B)
-from plasmapy.physics.parameters import (Hall_parameter, _grab_charge)
-from plasmapy.transport import (fundamental_electron_collision_freq,
-                                fundamental_ion_collision_freq)
-from plasmapy.transport.collisions import Coulomb_logarithm
+from plasmapy.formulary.physics.parameters import (Hall_parameter, _grab_charge)
+from plasmapy.formulary.transport import (fundamental_electron_collision_freq,
+                                          fundamental_ion_collision_freq)
+from plasmapy.formulary.transport import Coulomb_logarithm
 from plasmapy.utils import PhysicsError
